requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
polygon-api-client==0.2.11
requests==2.27.1
urllib3==1.26.9
websocket-client==1.3.2
websockets==10.3
Werkzeug==2.1.1
zipp==3.8.0
